+++ examples/variables.tf
@@ -50,7 +50,28 @@ variable "list-1" {
 }
 
 // A variable with underscores.
-variable "variable_with_underscores" {}
+variable "input_with_underscores" {}
+
+// A variable with pipe in the description
+variable "input-with-pipe" {
+  description = "It includes v1 | v2 | v3"
+  default     = "v1"
+}
+
+variable "input-with-code-block" {
+  description = <<EOD
+This is a complicated one. We need a newline.  
+And an example in a code block
+```
+default     = [
+  "machine rack01:neptune"
+]
+```
+EOD
+  default     = [
+    "name rack:location"
+  ]
+}
 
 variable "long_type" {
   type = object({